selectors added  for round butons and different gestures from users. Button sizes adjusted with rest of the GUI, not so long buttons anymore
selectors added  for round butons and different gestures from users.
Button sizes adjusted with rest of the GUI, not so long buttons anymore.
Default text added for security quiz and Statistics view if not user not
logged in. new Strings added in Strings.xml. Muses text added with Muses
icon on application bar. New icons added for Muses statuses with green
and red cricle showing Muses is ONLINE or OFFLINE.